requirements.txt
fastapi
uvicorn
python-multipart
pypdf
python-dotenv
chromadb==0.5.23
langchain==0.3.25
langchain-community==0.3.24
langchain-text-splitters==0.3.5
langchain-google-genai==2.1.4
google-generativeai
groq